feat: moves modules to folders
