+++ src/components/api/enum.ts
@@ -11,14 +11,3 @@ export enum Method {
   PUT = "PUT",
   DELETE = "DELETE",
 }
-
-export enum SortWinners {
-  Id = "id",
-  Wins = "wins",
-  Time = "time",
-}
-
-export enum Order {
-  Asc = "ASC",
-  Desc = "Desc",
-}
