+++ src/constants/enum.ts
@@ -1,0 +1,10 @@
+export enum SortWinners {
+  Id = "id",
+  Wins = "wins",
+  Time = "time",
+}
+
+export enum Order {
+  Asc = "ASC",
+  Desc = "Desc",
+}
