+++ src/constants/types.ts
@@ -1,7 +1,3 @@
-export enum NodeType {
-  ELEMENT_NODE = 1,
-}
-
 export type CallBack = () => void;
 
 export type EngineStatus = "started" | "stopped" | "drive";
